academico_repo/requirements.txt
alembic==1.14.1
arrow==1.3.0
binaryornot==0.4.4
certifi==2024.12.14
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cookiecutter==2.6.0
greenlet==3.1.1
hupper==1.12.1
idna==3.10
Jinja2==3.1.5
Mako==1.3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
packaging==24.2
PasteDeploy==3.1.0
plaster==1.1.2
plaster-pastedeploy==1.0.1
psycopg2==2.9.10
psycopg2-binary==2.9.10
Pygments==2.19.1
PyJWT==2.10.1
pyramid==2.0
pyramid-jwt==1.6.1
pyramid-mako==1.1.0
pyramid-sqlalchemy==1.6
python-dateutil==2.9.0.post0
python-slugify==8.0.4
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
rich==13.9.4
setuptools==75.8.0
six==1.17.0
SQLAlchemy==2.0.37
text-unidecode==1.3
transaction==5.0
translationstring==1.4
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
urllib3==2.3.0
venusian==3.1.1
waitress==3.0.2
WebOb==1.8.9
zope.deprecation==5.1
zope.interface==7.2
zope.sqlalchemy==3.1
